+++ CHANGELOG.md
@@ -34,6 +34,7 @@
     - Add Dataset writeTo() method
     - Pipeline and Committee Machine are no longer Verbose
     - Remove K Best feature selector (special case of RFE)
+    - Changed Error Analysis metrics
 
 - 0.4.1
     - Optimized CART node splitting for low variance continuous features